refactor(findOrAdd): externalizing internal utils methods
